Changed index variable types to

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: mrahmat- <mrahmat-@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 11:37:39 by mrahmat-          #+#    #+#             */
-/*   Updated: 2024/05/08 10:35:22 by mrahmat-         ###   ########.fr       */
+/*   Updated: 2024/05/10 10:46:36 by mrahmat-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_bzero(void *str, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (n == 0)
